basic group chat and home page with groups; fix db and model types

+++ internal/model/message.go
@@ -7,9 +7,9 @@ import (
 const MessageContentMaxLength int = 8000
 
 type Message struct {
-	Id        uint      `db:"id"`
-	GroupId   uint      `db:"group_id"`
-	AuthorId  uint      `db:"author_id"`
+	Id        uint64    `db:"id"`
+	GroupId   uint64    `db:"group_id"`
+	AuthorId  uint64    `db:"author_id"`
 	Content   string    `db:"content"`
 	CreatedAt time.Time `db:"created_at"`
 }
