+++ internal/model/roles.go
@@ -1,7 +1,7 @@
 package model
 
 type Role struct {
-	GroupId  uint `db:"group_id"`
-	UserId   uint `db:"user_id"`
-	RightsId uint `db:"role_id"`
+	GroupId  uint64 `db:"group_id"`
+	UserId   uint64 `db:"user_id"`
+	RightsId uint32 `db:"role_id"`
 }
