+++ internal/model_request/group_delete.go
@@ -1,5 +1,5 @@
 package model_request
 
 type GroupDelete struct {
-	GroupId uint `uri:"group_id"`
+	GroupId uint64 `uri:"group_id"`
 }
